Added temp-template property
The temporary file name itself does not actually matter but clients using
the allocator may wish to select specific storage (partition) where the
temporary file should be created.

Also for consistency "file-size" property is readable. There's no point in
doing the same for "temp-template" property as the file name is deleted
from the file system immediately so it's useless to read it.