Allocating/deallocating disk space if supported
Currently the allocator doesn't reclaim deallocated file block to make it
reusable. This is for simplicity at this stage.

Give the simplicity described above, at least it's nice to reclaim the
actual disk space when freeing a memory block (if supported). This also
means that the allocation should guarantee that the disk space is available
if in the future implementation memory blocks are reusable.

If manipulating disk space is not supported then potential disk space
scarcity is left for mmap() to fail.

Replaced mkstemp() and unlink() with their g_* equivalents for portability.