filememallocator: added ability to reset the allocator
Still haven't tackled any strategy for reusing blocks so the simplest
solution for now seems to be to just reset the allocator.

This is not elegant though as currently the assumption is that the
allocator is potentially a global/shared resource.  So need to be careful
to reset the allocator in the right place.