gstfilememallocator: allocation shouldn't fail if fallocate is not supported
filememallocator attempts to use fallocate() if available.  This is to
pre-allocate file chunks and guarantee that the requested disk space is
available and reserved at the time of allocation.

Similarly when a memory block is deallocated, filememallocator uses
fallocate() to punch wholes in the file if this feature is available.  This
is to release the disk space if in the future the allocator provides some
reallocation schemes.

In the former case if fallocate() is available but the operation is not
supported, it always failed and returned NULL.  This functionality is
optional though and if fallocate() clearly indicates that the operation is
not supported (rather than some other error), the allocation should be
continued with the only risk of running out of disk space later on.

This was discovered on tmpfs on SL6 with 2.6.32-71.24.1.el6.i686 kernel.